Change content panel cards sizing and finish rest
I finally understood why my cards would overflow the grid tracks and
limits it would be contained within. It's because I didn't get that a
purcentage unit would be of the parent size yes, but that defining 100%
of it wouldn't take the remaining space but the exact same size as the
parent (so here with the title it overflows). Then I simply added the
trending section without too much trouble and that's it ! I'm happy with
the result. Cheers !